+++ source/extensions/lightspeed.trex.material_properties.shared.widget/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [2.0.3]
+## Changed
+- Increased delays and fixed bookeeping for tests 
+
 ## [2.0.2]
 ## Fixed
 - Fixed override functions to account for instance prims